+++ changelog.md
@@ -24,6 +24,7 @@ All notable changes to this project will be documented in this file.
 * Reimplement `Union` using
   <http://okmij.org/ftp/Haskell/extensible/OpenUnion51.hs> as a basis.
   [#14](https://github.com/IxpertaSolutions/freer-effects/issues/14)
+* Renamed `Teletype` example DSL to `Console`.
 
 ## [0.2.4.1] (November 25, 2016)